src/face_cut/requirements.txt
pillow
ydb